+++ StarterKit/ScreenshotInASheetView.swift
@@ -4,16 +4,13 @@ import AspirinShot
 
 @available(iOS 15.0, macOS 13.3, *)
 extension Screenshot {
-    static let titleAbove = Screenshot("Title Above", for: [.iPhone, .iPad(.landscape)])
+    static let inASheet = Screenshot("In a Sheet")
 }
 
 @available(iOS 15.0, macOS 13.3, *)
-struct TitleAboveScreenshotView: View {
+struct ScreenshotInASheetView: View {
     var body: some View {
-        // Play with .fit/.fill:
-        // * .fill for a zoomed in effect, partial view
-        // * .fit to see the entire view
-        ScreenshotWithTitleView(position: .top(0.2), screenshot: .fit(0.95, .top)) {
+        ScreenshotWithTitleView(position: .top(0.16), screenshot: .fill(0.8)) {
             Rectangle()
                 .foregroundStyle(.white)
                 .overlay {
@@ -24,18 +21,14 @@ struct TitleAboveScreenshotView: View {
                     }
                     .font(.largeTitle)
                 }
-                // Remove this if you don't want to present the view as a sheet
                 .screenshotAsSheet()
-                // Play with .fit/.fill: the best option depends on the device and its screen format
                 .productBezel(scaledTo: .fit, statusBar: .white)
-            
         } title: {
             /// You'll notice that the number of lines change between localizations.
-            AutoResizableText("You can't go wrong\nwith a **title above**")
+            AutoResizableText("Present **your view**\nin a **sheet**")
                 .multilineTextAlignment(.center)
                 .foregroundColor(.white)
                 .padding()
-                .padding(.horizontal)
         } background: {
             LinearGradient(
                 colors: [.yellow, .orange, .red],
@@ -47,17 +40,11 @@ struct TitleAboveScreenshotView: View {
 }
 
 @available(iOS 15.0, macOS 13.3, *)
-struct TitleAboveScreenshotView_Previews: PreviewProvider {
+struct ScreenshotInASheetView_Previews: PreviewProvider {
     static var previews: some View {
-        ScreenshotPreviews(for: [.iOS, .iOS(.landscape)]) {
-            TitleAboveScreenshotView()
+        ScreenshotPreviews(.inASheet) {
+            ScreenshotInASheetView()
         }
-
-        TitleAboveScreenshotView()
-            .screenshotPreview("en", displayName: "EN - 2 Lines")
-
-        TitleAboveScreenshotView()
-            .screenshotPreview("fr", displayName: "FR - 3 Lines")
     }
 }
 #endif