+++ StarterKit/ScreenshotsASApp.swift
@@ -10,24 +10,38 @@ struct ScreenshotsASApp_Previews: PreviewProvider {
                 // Portrait
                 .basic,
                 .framed,
+                .minimal,
+                .minimalStyle,
                 .titleAbove,
+                .titleAboveZoomedIn,
                 .titleBelow,
                 .uiKit,
+                .widgets,
                 // Landscape
                 .autoStatusBars,
                 .blackStatusBars,
                 .whiteStatusBars,
+                .autoBottomBars,
+                .blackBottomBars,
+                .whiteBottomBars,
             ]
         ) { screenshot in
             switch screenshot {
                 case .basic: BasicScreenshotView()
                 case .framed: FramedScreenshotView()
                 case .titleAbove: TitleAboveScreenshotView()
+                case .titleAboveZoomedIn: TitleAboveZoomedInScreenshotView()
                 case .titleBelow: TitleBelowScreenshotView()
                 case .uiKit: UIKitScreenshotView()
+                case .widgets: WidgetsScreenshotView()
+                case .minimal: MinimalScreenshotView()
+                case .minimalStyle: MinimalStyleScreenshotView()
                 case .autoStatusBars: AutoStatusBarsScreenshotView()
                 case .blackStatusBars: BlackStatusBarsScreenshotView()
                 case .whiteStatusBars: WhiteStatusBarsScreenshotView()
+                case .autoBottomBars: AutoBottomBarsScreenshotView()
+                case .blackBottomBars: BlackBottomBarsScreenshotView()
+                case .whiteBottomBars: WhiteBottomBarsScreenshotView()
                 default: EmptyView()
             }
         }